0b093e14079a...40b4ac92514c

+++ l2junity-gameserver/src/main/java/org/l2junity/gameserver/ai/AirShipAI.java
@@ -35,7 +35,7 @@ public class AirShipAI extends VehicleAI
 	}
 	
 	@Override
-	protected void moveTo(int x, int y, int z)
+	protected void moveTo(double x, double y, double z)
 	{
 		if (!_actor.isMovementDisabled())
 		{
