+++ l2junity-gameserver/src/main/java/org/l2junity/gameserver/model/zone/form/ZoneNPoly.java
@@ -21,7 +21,7 @@ package org.l2junity.gameserver.model.zone.form;
 import java.awt.Polygon;
 
 import org.l2junity.commons.util.Rnd;
-import org.l2junity.gameserver.GeoData;
+import org.l2junity.gameserver.geodata.GeoData;
 import org.l2junity.gameserver.model.Location;
 import org.l2junity.gameserver.model.itemcontainer.Inventory;
 import org.l2junity.gameserver.model.zone.L2ZoneForm;
@@ -51,19 +51,19 @@ public class ZoneNPoly extends L2ZoneForm
 	}
 	
 	@Override
-	public boolean isInsideZone(int x, int y, int z)
+	public boolean isInsideZone(double x, double y, double z)
 	{
 		return (_p.contains(x, y) && (z >= _z1) && (z <= _z2));
 	}
 	
 	@Override
-	public boolean intersectsRectangle(int ax1, int ax2, int ay1, int ay2)
+	public boolean intersectsRectangle(double ax1, double ax2, double ay1, double ay2)
 	{
 		return (_p.intersects(Math.min(ax1, ax2), Math.min(ay1, ay2), Math.abs(ax2 - ax1), Math.abs(ay2 - ay1)));
 	}
 	
 	@Override
-	public double getDistanceToZone(int x, int y)
+	public double getDistanceToZone(double x, double y)
 	{
 		int[] _x = _p.xpoints;
 		int[] _y = _p.ypoints;
@@ -95,7 +95,7 @@ public class ZoneNPoly extends L2ZoneForm
 	}
 	
 	@Override
-	public void visualizeZone(int z)
+	public void visualizeZone(double z)
 	{
 		int[] _x = _p.xpoints;
 		int[] _y = _p.ypoints;
